Update the location of the MIME cache
The XDG MIME specification[1] requires that software installations
create files in $XDG_DATA_DIRS/mime/packages and then run
update-mime-database, which will create a lot of files in
$XDG_DATA_DIRS/mime, notably the binary mime.cache.

In Clear, we don't want to store caches in /usr, so we make
/usr/share/mime be a symlink to /var/cache/mime and software installs
its XML files to /usr/share/mime-packages. In order to keep compliance
with the spec, we create a symlink to that in /var/cache/mime/packages
(so /usr/share/mime/packages will contain the packages).

Testing with Qt's QMimeDatabase class shows that it finds the same
content for both the binary cache and the source XML files.

[1] https://freedesktop.org/wiki/Specifications/shared-mime-info-spec/

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>